requirements.txt
fastapi
uvicorn[standard]
python-escpos
jinja2